split config into vars and lib paths, make VM pooling possible
